chore: clean

+++ core/message_manager.py
@@ -19,7 +19,7 @@ logger = get_logger("message_processor", "cyan")
 
 bot_config = global_config["bot_config"].get("bot")
 
-config_max_message_interval = int(bot_config.get("max_message_interval"))
+config_max_message_interval = float(bot_config.get("max_message_interval"))
 config_max_buffer_messages = int(bot_config.get("max_buffer_messages"))
 
 config_min_message_delay = float(bot_config.get("min_message_delay", "0.8"))
